Use Tap Area instead of Long Press
Remove Long-Presses as they have been to unreliable.
Instead when a user taps the upper half (roughly the
more indicator area) we will transition to the
multi device view.
The lower half will react as a regular on/off switch.
If a scene has just one device, the full tap area
will be used to trigger the scene.